Clear settings whenever we change servers
